Recommendation System/requirements.txt
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0